\name{estimate.theta}
\alias{estimate.theta}
\title{Estimate best estimate of theta.}
\usage{
  estimate.theta(in.data, n = max(in.data), v = TRUE)

  estimate.theta(in.data, n = max(in.data), v = TRUE)
}
\arguments{
  \item{in.data}{input data drawn from binomial
  distribution}

  \item{n}{number of subunits, default is max(in.data)}

  \item{in.data}{input data drawn from binomial
  distribution}

  \item{n}{number of subunits, default is max(in.data)}
}
\description{
  Returns the optimal point estimate of theta for the input
  data set. By default, theta is optimized with n set to
  the largest value of observed in the data.

  Returns the optimal point estimate of theta for the input
  data set. By default, theta is optimized with n set to
  the largest value of observed in the data.
}

