\name{gamma}
\alias{gamma}
\title{Calculate gamma}
\usage{
  gamma(input.data, Bayes = TRUE)
}
\arguments{
  \item{input.data}{the input data}

  \item{Bayes}{flag to specify whether Bayesian or
  frequentist method is to be used}
}
\description{
  For a given dataset, calculate the probability that the
  largest observations are artifactual, as described by eqn
  [10] in the main paper.
}

